+++ blog/src/main/java/com/michael/blog/service/PostServiceImpl/PostServiceImpl.java
@@ -7,6 +7,7 @@ import com.michael.blog.payload.response.PostResponse;
 import com.michael.blog.payload.response.PostResponsePagination;
 import com.michael.blog.repository.PostRepository;
 import com.michael.blog.service.PostService;
+import com.michael.blog.service.UserService;
 import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,18 +23,26 @@ import java.util.stream.Collectors;
 @Service
 public class PostServiceImpl implements PostService {
 
-    @Autowired
+
     private PostRepository postRepository;
+    private ModelMapper mapper;
+    private UserService userService;
 
     @Autowired
-    private ModelMapper mapper;
+    public PostServiceImpl(PostRepository postRepository, ModelMapper mapper, UserService userService) {
+        this.postRepository = postRepository;
+        this.mapper = mapper;
+        this.userService = userService;
+    }
 
     @Override
     public PostResponse createPost(PostRequest postRequest) {
         Post post = Post.builder()
+                .username(userService.getLoggedInUser().getUsername())
                 .title(postRequest.getTitle())
                 .description(postRequest.getDescription())
                 .content(postRequest.getContent())
+                .user(userService.getLoggedInUser())
                 .build();
         post = postRepository.save(post);
         return mapper.map(post, PostResponse.class);
@@ -63,6 +72,16 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
+    public List<PostResponse> getMyPosts() {
+
+        List<Post> posts = postRepository.getAllByUserId(userService.getLoggedInUser().getId());
+        return posts.stream()
+                .map(post->mapper.map(post, PostResponse.class))
+                .collect(Collectors.toList());
+    }
+
+
+    @Override
     public PostResponse getPostById(Long postId) {
         return mapper.map(getPostFromDB(postId), PostResponse.class);
     }
@@ -70,6 +89,9 @@ public class PostServiceImpl implements PostService {
     @Override
     public MessageResponse deletePost(Long postId) {
         Post post = getPostFromDB(postId);
+        if (!post.getUser().getId().equals(userService.getLoggedInUser().getId())){
+           throw new RuntimeException("This post doesn't belong to you, you can't delete it!");
+        }
         postRepository.delete(post);
         return new MessageResponse(String.format("Post with id: %s was deleted", postId));
     }
@@ -77,6 +99,9 @@ public class PostServiceImpl implements PostService {
     @Override
     public PostResponse updatePost(Long postId, PostRequest postRequest) {
         Post post = getPostFromDB(postId);
+        if (!post.getUser().getId().equals(userService.getLoggedInUser().getId())){
+            throw new RuntimeException("This post doesn't belong to you, you can't update it!");
+        }
         post.setTitle(postRequest.getTitle());
         post.setDescription(postRequest.getDescription());
         post.setContent(postRequest.getContent());
